First start of Data-Muddle
DataMuddle is for replacing and changing data so you can move it from
production to test environment and make it hard to figure out the
original data. The feel of the data should be the same.  I have added
Rhino-Etl references to it. Started on test data and first test cases.